requirements.txt
requests>=2.31.0
pandas>=1.3.0
numpy>=1.24.0
beautifulsoup4>=4.11.0
psycopg2-binary>=2.9.9
pg8000>=1.31.2
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pillow>=8.0.0

# Core app UI
streamlit>=1.28.0
plotly>=5.0.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
statsmodels>=0.14.0
joblib>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text processing and sentiment analysis
nltk>=3.8.1
textblob>=0.17.1
transformers>=4.30.0

# Backend integrations
supabase>=2.0.0
